create admin action, api, routes, db function
