+++ client/api/clients.js
@@ -33,11 +33,6 @@ export function updateClientDetails (clientDetails, token) {
     .catch(errorHandler('PATCH', '/api/v1/clients/:id/update'))
 }
 
-export function fetchClients () {
-  return request.get('/api/v1/clients')
-    .then((res) => res.body)
-}
-
 function errorHandler (method, route) {
   return (err) => {
     console.log(err)
