+++ server/db/admin.js
@@ -1,0 +1,10 @@
+const connection = require('./connection')
+
+module.exports = {
+  fetchAllClients
+}
+
+function fetchAllClients (db = connection) {
+  return db('clients')
+    .select()
+}
