+++ server/server.js
@@ -2,6 +2,7 @@ const path = require('path')
 const express = require('express')
 
 const clients = require('./routes/clients')
+const admin = require('./routes/admin')
 const publicRoutes = require('./routes/public')
 const protectedRoutes = require('./routes/protected')
 const privateRoutes = require('./routes/private')
@@ -12,6 +13,7 @@ server.use(express.json())
 server.use(express.static(path.join(__dirname, 'public')))
 
 server.use('/api/v1/clients', clients)
+server.use('/api/v1/admin', admin)
 server.use('/api/v1/public', publicRoutes)
 server.use('/api/v1/protected', protectedRoutes)
 server.use('/api/v1/private', privateRoutes)
